+++ lib/to_source/emitter/back_ref.rb
@@ -1,0 +1,23 @@
+module ToSource
+  class Emitter
+
+    # Emitter for various access nodes
+    class BackRef < self
+
+      handle(Rubinius::AST::BackRef)
+
+    private
+
+      # Perform dispatch
+      #
+      # @return [undefined]
+      #
+      # @api private
+      # 
+      def dispatch
+        emit("$#{node.kind}")
+      end
+
+    end
+  end
+end