+++ lib/to_source/emitter/element_reference.rb
@@ -16,7 +16,7 @@ module ToSource
       def dispatch
         visit(receiver)
         emit('[')
-        visit(node.arguments.array.first)
+        visit(node.arguments)
         emit(']')
       end