+++ lib/to_source/emitter/multiple_assignment.rb
@@ -1,9 +1,10 @@
 module ToSource
   class Emitter
-    # Emitter for multiple assignment nodes
-    class MultipleAssignment < self
 
-      handle(Rubinius::AST::MultipleAssignment)
+    class Splat < self
+      handle(Rubinius::AST::SplatArray)
+      handle(Rubinius::AST::SplatAssignment)
+      handle(Rubinius::AST::SplatWrapped)
 
     private
 
@@ -14,9 +15,34 @@ module ToSource
       # @api private
       #
       def dispatch
+        emit('*')
+        visit(node.value)
+      end
+    end
+
+    # Emitter for multiple assignment nodes
+    class MultipleAssignment < self
+
+      handle(Rubinius::AST::MultipleAssignment)
+
+    private
+
+      delegate(:right, :splat, :left)
+      predicate(:right, :splat, :left)
+
+      # Perform dispatch
+      #
+      # @return [undefined]
+      #
+      # @api private
+      #
+      def dispatch
         emit_left
-        emit(' = ')
-        emit_right
+        emit_splat
+        if right?
+          emit(' = ')
+          emit_right 
+        end
       end
 
       # Emit left
@@ -26,16 +52,30 @@ module ToSource
       # @api private
       #
       def emit_left
-        run(Util::DelimitedBody, node.left.body)
+        return unless left?
+        run(Util::DelimitedBody, left.body)
       end
 
-      # Emit left
+      # Emit splat
+      #
+      # @return [undefined]
+      #
+      # @api private
+      #
+      def emit_splat
+        return unless splat?
+        emit(', ') if left?
+        visit(splat)
+      end
+
+      # Emit right
       #
       # @return [undefined]
       #
       # @api private
       #
       def emit_right
+        return unless right?
         right = node.right
         if right.kind_of?(Rubinius::AST::ArrayLiteral)
           run(Util::DelimitedBody, right.body)